+++ kork-retrofit/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.netflix.spinnaker.kork.web.serviceclient.RetrofitServiceFactoryAutoConfiguration
+  com.netflix.spinnaker.kork.retrofit.RetrofitServiceFactoryAutoConfiguration
